meta-linaro: ganglia: add monitoring utilities to system startup
Also added sample configuration files.  The multicast example would work
well on a network with suitable DNS resolution.  The default
configuration that is used is a copy of the host-example.conf and that
directs all monitoring information to a well-known IP address which
works well on the ARMv8 Foundation Model.

Change-Id: I5274b722245cbc2476b08c293a5b25d2cd25e36b
Signed-off-by: Andrew McDermott <andrew.mcdermott@linaro.org>