libhugetlbsfs git: replace AUTOREV with a fixed SRCREV
Using AUTOREV is bad for the following reasons:

* Makes bisecting bugs impossible since it will always use the latest rev, so rebuilding old releases that include or depend on one of these recipes is impossible
* Since the maintainance last weekend git.linaro.org has been flakey causing not only our CI loop to fail miserably, but it also stopped non-linaro users of meta-linaro from using it.
* It introduces changes during the freeze period of the cycle

It also causes problems when the patches in the recipe get applied to the git tree, there will be a window where the patch is in both places and the build fails. That happened last month.

This recipe can be manually or automatically updated at the start of the cycle, but shouldn't change during the freeze period at the end of each cycle.

Change-Id: I9f8a3d67c96f327703f153d442bffdb4e1a40e2e
Signed-off-by: Koen Kooi <koen.kooi@linaro.org>