bootwrapper: fix the configure error
Since kernel commit ca5b34100c57 ("dts, arm64: Move dts files to vendor
subdirs"), the default dtb directory in bootwrapper is not matched with
the kernel and it would provoke the following error when checking the
dtb in configure:
  checking for build/arch/arm64/boot/dts/rtsm_ve-aemv8a.dtb... no
  configure: error: No such file or directory: build/arch/arm64/boot/dts/rtsm_ve-aemv8a.dtb
  WARNING: temp/run.do_compile.5170:1 exit 1 from

The latest bootwrapper adds a "--with-dtb" to override the default dtb
in configure.ac. So update the bootwrapper to the latest version and
use the "--with-dtb" option to workaround this build issue.

Change-Id: Ic664e959ce8adb742dfce5b12b1981fbb3995e56
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Koen Kooi <koen.kooi@linaro.org>