odp: update recipe to use Release 1.0 source
Patch 1 of 2

Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>

odp: fix compiler warning/error re. use of _BSD_SOURCE

Patch 2 of 2

This patch allows compilation of the ODP v1.0 sources to complete.
It is added to the OE recipe for odp since that recipe is pinned to
ODP release version 1.0, which requires the patch.  Details of the
problem are found in the patch file itself and in Linaro BUG #1312

Upstream status:  The patch added here has been accepted into the
                  upstream ODP repository, but is nonetheless still
                  required for this recipe since the source revision
                  upon which the recipe is based lacks the patch.

Change-Id: I678d84a418473cd47b8503603b1320806be24db0
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>