lng-network-config: fix 'S' path to sources
The path to the source files for this package is in a subdirectory 'files'
off of the build directory - rather than being in a subdirectory named
via the package base name and  version.
Previous commit 8ab9cdc17a4e865cd9cf45edf7a0c167644d08d1 assumed the latter
source directory naming convention and did not resolve the failed path.
This commit resolves the problem so the udev rule file can be found.

Change-Id: I7484ffeb7e40b15f53723be6a45e2f7063374da9
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>