lng-network-config: set 'S' to correct source path
The source files for this recipe are unpacked into ${WORKDIR},
rather than being in a subdirectory named via the package base name
and version... and the separate subdir does not exist for that reason.
Previous commit 8ab9cdc17a4e865cd9cf45edf7a0c167644d08d1 assumed the
existence of the (normally present) source subdirectory.
This commit resolves the problem so an existing path is used for 'S'.

Change-Id: I88f6eabf286e940edb790885cc8232927729f05d
Signed-off-by: Gary S. Robertson <gary.robertson@linaro.org>